requirements.txt
fastapi
nats-python
nats-py
requests



#python.exe -m pip install --upgrade pip